split context into another file
